added simple instructions for import/export db to gui-guide added some gui problems to problems page
in defaultdb:
describe more about which traits may need adjusting
added section on building the extra packages

fix a couple of errors, omissions in abstract

in quickstart-vbox:
update netinst iso version to 501
paella-user-quick-setup script now does more, and deprecates some of the 
instructions.  Try to let the reader know about this, and leave those instructions 
intact as a reference. (although this makes the page a little uglier)

reorder index page